+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.1.1 (2016/3/15)
+
+Features:
+
+- New subcommand `set-role` to add or update "role" of an Instance or an AMI.
+- New subcommand `set-group` to add or update "group" of an Instance or an AMI.
+
+Improvements:
+
+- `list`, `list-ami` ... Sort by "name" or "disp_name"
+
 ## 0.1.0 (2016/3/13)
 
 First release.